feat: InputView의 메서드 수정 & scanner 입력받도록 수정
